+++ packages/share_plus/share_plus_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.3
+
+- Update dependencies
+
 ## 3.0.2
 
 - Gracefully fall back from `shareWithResult` to regular `share` methods on unsupported platforms